requirements.txt
openai
langchain
langchain_openai
langchain_community
langchain_experimental
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
numpy
pandas
chromadb
streamlit
Wikipedia
pypdf
docx2txt
